==> src/01_load_data.ipynb <==
import pandas as pd
import os
import json
from pandas import json_normalize
from ast import literal_eval

pd.set_option('display.max_columns', None)

Download the CCV data from https://github.com/versotym/corpusCzechVerse and save them to your local folder.

# EDIT path to your data folder 
BASEDIR = '../../corpusCzechVerse/ccv/'

# Load the data

file_names = sorted(os.listdir(BASEDIR))

# number of books
len(file_names)

%%time

no_poems = 0

for file_name in file_names:
    book = json_normalize(json.load(open(BASEDIR + file_name)))
    no_poems += len(book)
    
print('Number of poems:', no_poems)

%%time
books = []

for i in range(0, 600):
    book = json_normalize(json.load(open(BASEDIR + file_names[i])))
    books.append(book)

len(books)

df = pd.concat(books)

len(df)

df.head()

poem_1 = df.body[0]

len(poem_1)

# poem_1
for stanza in poem_1:
    for line in stanza:
        print(line['text'])
    print('')  # end of stanza

# save the loaded books into a single file
df.to_csv('0001-0677.csv', index=False)

# Read complete df with all books

import pandas as pd

df = pd.read_csv('../data/books.csv', low_memory=False)

df.head()

len(df)

df.dtypes

# Store x% sample for testing

x = 0.2

books = df.sample(frac=x, random_state=42, ignore_index=True)

books

books.to_csv('../data/books_frac_0.2.csv', index=False)

# Store just poem id and text (for embeddings)

len(df)

df['book_poem_id'] = df['book_id'].astype(str) + '_' + df['poem_id']

df = df[['book_poem_id', 'body']]

df.head()

%%time

texts = []

for index, p in df.iterrows():
    text = ''
    poem = literal_eval(p['body'])
    for stanza in poem:
        for line in stanza:
            text += line['text']
            text += ' '
    
    if index % 10000 == 0:
        print(index)    
    
    texts.append(text)

len(texts)

df['text'] = texts

texts_df = df.drop(columns=['body'])

texts_df

texts_df['text'][13]

texts_df.to_csv('../data/texts.csv', index=False)

# Store body with book_poem_id

df['book_poem_id'] = df['book_id'].astype(str) + '_' + df['poem_id']

df = df[['book_poem_id', 'body']]

df.to_csv('../data/bodies.csv', index=False)

# Store book_poem_id, lemmas and their POS tags

df['book_poem_id'] = df['book_id'].astype(str) + '_' + df['poem_id']

df = df[['book_poem_id', 'body']]

lemmas_col = []

# extract lemmas and pos tags
for index, poem in df.iterrows():
    if index % 100 == 0:
        print(index, end=' ')
        
    lemmas_poem = []
    body = literal_eval(poem['body'])
    for stanza in body:
        for line in stanza:
            for word in line['words']:
                # if POS tag is Noun, Adjective or Verb
                pos = word['morph'][0]
                lemmas_poem.append((word['lemma'], pos))
    lemmas_col.append(lemmas_poem)

df['lemmas'] = lemmas_col

df = df[['book_poem_id', 'lemmas']]

df.head(3)

df.shape

df.to_csv('../data/lemmas.csv', index=False)
