==> src/04_create_embeddings.ipynb <==
# Embeddings

import pandas as pd
import string

texts_df = pd.read_csv('../data/texts.csv')

# Preprocess texts

def preprocess(text):
    # convert input text to lower case.
    text = text.lower()
    # remove punctuations from string.
    translator = str.maketrans(string.punctuation, ' '*len(string.punctuation))
    text = text.translate(translator)
    # replace all whitespaces by a single space
    text = ' '.join(text.split())
    
    return text

preprocess(texts[0])

texts_df['text_preprocessed'] = texts_df['text'].apply(lambda x: preprocess(x))

# add text length info
texts_df['n_words'] = texts_df['text_preprocessed'].apply(lambda x: len(x.split()))

texts_df.head()

texts_df.text_len.describe()

# Create embeddings for poem texts

from sentence_transformers import SentenceTransformer
model = SentenceTransformer('sentence-transformers/distiluse-base-multilingual-cased-v2')

#Our sentences we like to encode
sentences = texts_df.text_preprocessed.to_numpy()

#Sentences are encoded by calling model.encode()
embeddings = model.encode(sentences)

len(embeddings)

embeddings.shape

embeddings_list = embeddings.tolist()

len(embeddings_list)

texts_df['embedding'] = embeddings_list

embeddings_df = texts_df[['poem_id', 'embedding']]

embeddings_df.head()

embeddings_df.to_csv('../data/embeddings.csv', index=False)
