==> src/11_supervised_2.ipynb <==
# Supervised Learning - part 2

- Only the best model, all data.

import pandas as pd
import numpy as np
from ast import literal_eval
from sklearn.feature_extraction.text import CountVectorizer
from wordcloud import WordCloud
import matplotlib.pyplot as plt

from gensim.corpora import Dictionary
from gensim.models import CoherenceModel

from sklearn import naive_bayes, svm
from sklearn.preprocessing import LabelEncoder
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer, HashingVectorizer
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, f1_score, confusion_matrix, classification_report, make_scorer
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.model_selection import GridSearchCV, KFold, cross_val_score
from imblearn.over_sampling import RandomOverSampler

from sklearn.naive_bayes import GaussianNB
from sklearn.neighbors import KNeighborsClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.linear_model import LogisticRegression

import warnings
warnings.filterwarnings('ignore')

RANDOM_STATE = 42

lemmas = pd.read_csv('../data/lemmas.csv')
lemmas

annotations = pd.read_csv('../data/poem_value_save.csv')
annotations

annotations_1 = annotations[annotations.value == 1].reset_index(drop=True)
annotations_1

annotations_lemmas = pd.merge(lemmas, annotations_1, on='book_poem_id', how='left')
annotations_lemmas

# Filter the data

- Filter out all POS but Nouns (N), Adjectives (A) and Verbs (V)

- Filter out frequent (stop)words that does not carry any extra semantic information

stopwords = ['mít', 'jít', 'být', 'dát', 'moci', 'chtít']

%%time

lemmas_filtered_column = []

# extract lemmas with N, A or V POS tag
for index, poem in annotations_lemmas.iterrows():
    if index % 1000 == 0:
        print(index, end=' ')
    
    lemmas_filtered_poem = []
    lemmas_poem = literal_eval(poem['lemmas'])
    for lemma_pos in lemmas_poem:
        lemma, pos = lemma_pos
        if pos in ['N', 'A', 'V'] and lemma not in stopwords:
            lemmas_filtered_poem.append(lemma)
    lemmas_filtered_column.append(lemmas_filtered_poem)

annotations_lemmas['lemmas_filtered'] = lemmas_filtered_column
annotations_lemmas

def c_tf_idf(documents, m, ngram_range=(1, 1)):    
    count = CountVectorizer(ngram_range=ngram_range, stop_words="english").fit(documents)
    t = count.transform(documents).toarray()
    w = t.sum(axis=1)
    tf = np.divide(t.T, w)
    sum_t = t.sum(axis=0)
    idf = np.log(np.divide(m, sum_t)).reshape(-1, 1)
    tf_idf = np.multiply(tf, idf)
    
    return tf_idf, count


def extract_top_n_words_per_cluster(tf_idf, count, docs_per_cluster, groupby, n=20):
    words = count.get_feature_names_out()
    labels = list(docs_per_cluster[groupby])
    tf_idf_transposed = tf_idf.T
    indices = tf_idf_transposed.argsort()[:, -n:]
    top_n_words = {label: [(words[j], tf_idf_transposed[i][j]) for j in indices[i]][::-1] for i, label in enumerate(labels)}
    return top_n_words


def extract_topic_sizes(df, groupby):
    topic_sizes = (df.groupby([groupby])
                     .preprocessed_text
                     .count()
                     .reset_index()
                     .rename({"preprocessed_text": "size"}, axis='columns')
                     .sort_values("size", ascending=False))
    return topic_sizes

annotations_lemmas = annotations_lemmas.drop(['name'], axis=1)

annotations_lemmas.head(1)

# TFIDF classification

data = pd.DataFrame()
data = annotations_lemmas[['book_poem_id', 'id_topic', 'lemmas_filtered']]
data['preprocessed_text'] = annotations_lemmas['lemmas_filtered'].apply(lambda x: ' '.join(x))
data

X_train2 = data[data['id_topic'].notnull()]
X_rest = data[~data['id_topic'].notnull()]

X = X_train2[['preprocessed_text']].copy()
y = X_train2[['id_topic']].copy()

X_rest2 = X_rest[['preprocessed_text']].copy()
y_rest2 = X_rest[['id_topic']].copy()

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42, stratify=y)

print(len(X_train), len(X_test))

y.value_counts()

# Scores of annotated data

X_train2

# Create the library
min_frequency = 20
dictionary = Dictionary(X_train2['lemmas_filtered'])
# Filter out words
dictionary.filter_extremes(no_below=min_frequency)

from collections import Counter

# create word frequency count
word_counts = Counter(word for row in X_train2['lemmas_filtered'] for word in row)

# filter lists based on word frequency count
X_train2['lemmas_filtered_a'] = X_train2['lemmas_filtered'].apply(lambda x: [word for word in x if word_counts[word] >= 20])

X_train2['preprocessed_text'] = X_train2['lemmas_filtered_a'].apply(lambda x: ' '.join(x))

# merge documents in cluster into one document
docs_per_cluster = X_train2.groupby(['id_topic'], as_index = False).agg({'preprocessed_text': ' '.join})

docs_per_cluster

tf_idf, count = c_tf_idf(docs_per_cluster.preprocessed_text.values, m=len(X_train2))

top_n_words = extract_top_n_words_per_cluster(tf_idf, count, docs_per_cluster, 'id_topic', n=10)

topic_sizes = extract_topic_sizes(X_train2, 'id_topic')
topic_sizes

topics = []

for i in top_n_words:
    words_topic = []
    for word in top_n_words[i]:
        words_topic.append(word[0])
    topics.append(words_topic)

cm = CoherenceModel(topics=topics, topn=10, dictionary=dictionary, texts=X_train2['lemmas_filtered_a'], coherence='c_v')
coherence = cm.get_coherence()
coherence

from utils import topic_diversity

topic_diversity(topics, 10)

top_n_words

# Oversampling

# Apply random oversampling to the minority class
from imblearn.over_sampling import RandomOverSampler

ros = RandomOverSampler(random_state=RANDOM_STATE)
X_train_sampled, y_train_sampled = ros.fit_resample(X_train, y_train)

y_train_sampled.value_counts()

# Encode target variable

X_train = X_train_sampled
y_train = y_train_sampled

# Encoding of target variable - convert text labels into numbers
encoder = LabelEncoder()
encoder_fitted = encoder.fit(y_train)

y_train = encoder_fitted.transform(y_train)
y_test = encoder_fitted.transform(y_test)

encoder_fitted.classes_

encoder_fitted.transform([ 1,  2,  3,  4,  9, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23,
       24, 25, 26, 27, 28, 29, 30])

# Vectorization

vectorizer_tfidf = TfidfVectorizer()  # good for SVM
#vectorizer_count = CountVectorizer()  # good for Naive Bayes

vectorizer_fitted_tfidf = vectorizer_tfidf.fit(X_rest2['preprocessed_text'])
#vectorizer_fitted_count = vectorizer_count.fit(X_train['preprocessed_text'])

train_X_encodings_tfidf = vectorizer_fitted_tfidf.transform(X_train['preprocessed_text'])
test_X_encodings_tfidf = vectorizer_fitted_tfidf.transform(X_test['preprocessed_text'])
rest_X_encodings_tfidf = vectorizer_fitted_tfidf.transform(X_rest2['preprocessed_text'])

#train_X_encodings_count = vectorizer_fitted_count.transform(X_train['preprocessed_text'])
#test_X_encodings_count = vectorizer_fitted_count.transform(X_test['preprocessed_text'])

train_X_encodings_tfidf

#train_X_embeddings_word = X_train['embedding_word'].to_list()
#test_X_embeddings_word = X_test['embedding_word'].to_list()

# ML

f1_macro_scorer = make_scorer(f1_score, average='macro')

scoring = 'accuracy'

# Create k-fold cross-validation object
kfold = KFold(n_splits=5, shuffle=True, random_state=RANDOM_STATE)

# The best model

# MANUALLY pick the best performing model
model_svm = MLPClassifier(activation= 'logistic', alpha=0.01, random_state=RANDOM_STATE) #probability=True
# train your model on the entire training set
model_svm.fit(train_X_encodings_tfidf, y_train)

# predict labels for the test set
y_pred = model_svm.predict(test_X_encodings_tfidf)

#print(classification_report(y_test, y_pred, labels=encoder_fitted.classes_))

accuracy_score(y_true=y_test, y_pred=y_pred)

f1_score(y_true=y_test, y_pred=y_pred, average='macro')

# on all set
y_pred_rest = model_svm.predict(rest_X_encodings_tfidf)

y_pred_rest

X_rest['predicted_label'] = y_pred_rest

X_rest

# Create the dictionary
min_frequency = 10
dictionary = Dictionary(X_rest['lemmas_filtered'])
# Filter out words
dictionary.filter_extremes(no_below=min_frequency)

from collections import Counter

# create word frequency count
word_counts = Counter(word for row in X_rest['lemmas_filtered'] for word in row)

# filter lists based on word frequency count
X_rest['lemmas_filtered_a'] = X_rest['lemmas_filtered'].apply(lambda x: [word for word in x if word_counts[word] >= 20])

X_rest['lemmas_filtered_a'] = X_rest['lemmas_filtered_a'].apply(lambda x: ' '.join(x))

mapping = dict(zip(encoder_fitted.transform(encoder_fitted.classes_), encoder_fitted.classes_))

X_rest.predicted_label.replace(mapping, inplace=True)

X_rest

# merge documents in cluster into one document
docs_per_cluster = X_rest.groupby(['predicted_label'], as_index = False).agg({'preprocessed_text': ' '.join})
docs_per_cluster

tf_idf, count = c_tf_idf(docs_per_cluster.preprocessed_text.values, m=len(X_rest))

top_n_words = extract_top_n_words_per_cluster(tf_idf, count, docs_per_cluster, 'predicted_label', n=10)

topic_sizes = extract_topic_sizes(X_rest, 'predicted_label')
topic_sizes

top_n_words

df_topics = pd.read_csv('../data/list_of_topics.csv')
df_topics

#for i in [k for k in top_n_words]:
#    print(i, df_topics[df_topics.id_topic == i].reset_index(drop=True).name[0])
#    word_cloud(top_n_words, i)

topics = []

for i in top_n_words:
    words_topic = []
    for word in top_n_words[i]:
        words_topic.append(word[0])
    topics.append(words_topic)
    
topics[2]

cm = CoherenceModel(topics=topics, topn=10, dictionary=dictionary, texts=X_rest['lemmas_filtered'], coherence='c_v')
coherence = cm.get_coherence()
coherence

topic_diversity(topics, 10)
